Adding the first version of the readobj library.
The readobj library is the core of this project.  It defines a Ruby object
that represents one object file (ObjFile); it currently takes a filename in its
constructor and parses that file, breaking it into tables for the
segments, symbols and relocations, and stores them in tables which are
accessible to the client.  It does not yet provide a way to write an
object file back to disk.

The overall concept is that a linker client will create one
ObjFile per in object file on disk, and compose new ObjFiles in the course
of 'linking'.

Note that the code in this file is closely adapted from John R. Levine's
Perl-based implementation on the web.

This includes some test code at the bottom; this code loads the file
ch4main.lk, creates an ObjFile, and prints info from the ObjFile to
screen.