Reserve a null entry in the segment and symbol tables.
I realized there was a bug in previous implementations of ObjFile: the
spec calls for all segment and symbol numbers to start at 1, not 0, with
an implied 'null' entry at location zero in the tables.  This was how
Levine's demonstration code was implemented, and I chose to imitate that
method.